Move Well migrations to own file
